Move get_wavelengths() to the instrument.py module, since it may apply to any mode. Move get_bin_edges to binning.py.
