+++ maven_iuvs/binning.py
@@ -1,4 +1,46 @@
 import numpy as np
+from maven_iuvs.instrument import get_wavelengths
+
+def get_bin_edges(light_fits):
+    """
+    Wavelengths as defined in the fits files are defined for the bin centers.
+    This function will calculate where the bin edges should be, since the 
+    recorded bin edges in the files are all for the standard resolution mode, 
+    and not able to be applied to echelle.
+    TODO: The method used here is probably "good enough" but could be improved.
+
+    Parameters:
+    -----------
+    light_fits : astropy.io.fits instance
+                 File with light observation
+    Returns:
+    -----------
+    edges : array
+            Defines the edges of the bins for the wavelengths, so we can calculate the flux
+            to assign to the bins.
+    """    
+
+    # Grab the wavelengths 
+    wavelengths = get_wavelengths(light_fits)
+
+    # First calculate the differences between all points x
+    dlambda = np.diff(wavelengths)
+    
+    # There will be one more bin edge than x points
+    edges = np.zeros(len(wavelengths) + 1)
+
+    # Handle the left edge
+    edges[0] = wavelengths[0] - (dlambda[0] / 2) 
+
+    # inner elements
+    for i in range(1, len(edges)-1):
+        edges[i] = wavelengths[i] - dlambda[i-1] / 2
+
+    # And the right edge
+    edges[-1] = wavelengths[-1] + dlambda[-1] / 2
+    
+    return edges
+
 
 def get_binning_scheme(hdul):
     """
